event-filtering-bag-processing/requirements.txt
matplotlib
numpy
tqdm
rosbag
imageio
Pillow